Added alerts for review deletion and addition

+++ src/components/UserReviewCard.tsx
@@ -12,16 +12,48 @@ export default function UserReviewCard({ review, reviews, setFilteredReviews, fi
 
 
 
-  const handleDeleteReview = async () => {    
-    setFilteredReviews(filteredReviews.filter((review) => review.review_id !== review_id))
-    const { data, error } = await supabase
-    .from(`${table}`)
-    .delete()
-    .eq('review_id', review_id);
-    if (error) {
-      Alert.alert('Something went wrong. Please try again.')
-      setFilteredReviews(reviews)
-    }
+  // const handleDeleteReview = async () => {    
+  //   setFilteredReviews(filteredReviews.filter((review) => review.review_id !== review_id))
+  //   const { data, error } = await supabase
+  //   .from(`${table}`)
+  //   .delete()
+  //   .eq('review_id', review_id);
+  //   if (error) {
+  //     Alert.alert('Something went wrong. Please try again.')
+  //     setFilteredReviews(reviews)
+  //   }
+  // }
+
+
+  const handleDeleteReview = async () => {  
+    
+    Alert.alert(
+      "Confirm Deletion",
+      "Are you sure you want to delete this review?",
+      [
+        {
+          text: "Cancel",
+          style: "cancel",
+        },
+        {
+          text: "Delete",
+          onPress: async () => {
+            setFilteredReviews(filteredReviews.filter((review) => review.review_id !== review_id))
+            const { data, error } = await supabase
+            .from(`${table}`)
+            .delete()
+            .eq('review_id', review_id);
+
+            if (error) {
+              Alert.alert('Something went wrong. Please try again.')
+              setFilteredReviews(reviews)
+            } 
+          },
+          style: "destructive",
+        },
+      ],
+      { cancelable: false }
+    );
   }
 
   const handleEditReview = async () => {
